Extracted the war into two new projects: App and Jar. The war and jar will product a war file and jar file respectively and the app project will contain the same classes that war project contained previously.
